judged questionlyx/a/a

+++ lyx/a/a.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 #define MAXN 1000020
 int ap[8080]={0};
-vector<int8_t> data[MAXN];
+vector<int> data[MAXN];
 int n,q;
 ifstream fin("a.in");
 ofstream fout("a.out");
